refactor: consolidate precommit and autocommit functions into tool files
- Move use_checks() from util/context.py to tools/precommit.py
- Move run_precommit_checks() from util/context.py to tools/precommit.py
- Move autocommit() from util/context.py to tools/autocommit.py
- Update gather_fresh_context() to import run_precommit_checks from precommit tool
- Update test imports to use new locations
- Remove unused time import from context.py

This continues the refactoring pattern established in the autocompact tool,
consolidating tool-specific functionality into the tool files themselves
rather than keeping them in util modules.
