+++ gptme/util/context.py
@@ -5,7 +5,6 @@ import os
 import re
 import shutil
 import subprocess
-import time
 import urllib
 import urllib.parse
 from collections import Counter
@@ -40,43 +39,6 @@ def use_fresh_context() -> bool:
     """
     flag: str = get_config().get_env("GPTME_FRESH", "")  # type: ignore
     return flag.lower() in ("1", "true", "yes")
-
-
-def use_checks() -> bool:
-    """Check if pre-commit checks are enabled.
-
-    Pre-commit checks are enabled when either:
-    1. GPTME_CHECK=true is set explicitly, or
-    2. A .pre-commit-config.yaml file exists in any parent directory
-
-    Any issues found are included in the context, helping catch and fix code quality
-    issues before the user continues the conversation.
-    """
-    flag: str = get_config().get_env("GPTME_CHECK", "")  # type: ignore
-    explicit_enabled = flag.lower() in ("1", "true", "yes")
-    explicit_disabled = flag.lower() in ("0", "false", "no")
-    if explicit_disabled:
-        return False
-
-    # Check for .pre-commit-config.yaml in any parent directory
-    has_config = any(
-        parent.joinpath(".pre-commit-config.yaml").exists()
-        for parent in [Path.cwd(), *Path.cwd().parents]
-    )
-
-    if explicit_enabled and not has_config:
-        logger.warning(
-            "GPTME_CHECK is enabled but no .pre-commit-config.yaml found in any parent directory"
-        )
-
-    enabled = explicit_enabled or has_config
-
-    # Check for pre-commit availability
-    if enabled and not shutil.which("pre-commit"):
-        logger.warning("pre-commit not found, disabling pre-commit checks")
-        return False
-
-    return enabled
 
 
 def file_to_display_path(f: Path, workspace: Path | None = None) -> Path:
@@ -260,6 +222,8 @@ def gather_fresh_context(
 
     # Add pre-commit check results if there are issues
     if precommit:
+        from ..tools.precommit import run_precommit_checks
+
         success, precommit_output = run_precommit_checks()
         if not success and precommit_output:
             sections.append(precommit_output)
@@ -315,132 +279,6 @@ def get_changed_files() -> list[Path]:
     except subprocess.CalledProcessError as e:
         logger.debug(f"Error getting git diff files: {e}")
         return []
-
-
-def run_precommit_checks() -> tuple[bool, str | None]:
-    """Run pre-commit checks on modified files and return output if there are issues.
-
-    Pre-commit checks will run if either:
-    1. GPTME_CHECK=true is set explicitly, or
-    2. A .pre-commit-config.yaml file exists in any parent directory
-
-    Returns:
-        A tuple (True, None) if no issues found,
-        or (False, output) if issues found,
-        or (False, None) if interrupted.
-        If pre-commit checks are not enabled, returns (False, None).
-    """
-    if not use_checks():
-        logger.debug("Pre-commit checks not enabled")
-        return False, None
-
-    # cmd = "pre-commit run --files $(git ls-files -m)"
-    cmd = "pre-commit run --all-files"
-    start_time = time.monotonic()
-    logger.info(f"Running pre-commit checks: {cmd}")
-    try:
-        subprocess.run(cmd, shell=True, capture_output=True, text=True, check=True)
-        return True, None  # No issues found
-    except subprocess.CalledProcessError as e:
-        # if exit code is 130, it means the user interrupted the process
-        if e.returncode == 130:
-            logger.info("Pre-commit checks interrupted by user")
-            return False, None
-        # If no pre-commit config found
-        # Can happen in nested git repos, since we check parent dirs but pre-commit only checks the current repo.
-        if ".pre-commit-config.yaml is not a file" in e.stdout:
-            return False, None
-
-        logger.error(f"Pre-commit checks failed: {e}")
-        output = "Pre-commit checks failed\n\n"
-
-        # Add stdout if present
-        if e.stdout.strip():
-            output += md_codeblock("stdout", e.stdout.rstrip()) + "\n\n"
-
-        # Add stderr if present
-        if e.stderr.strip():
-            output += md_codeblock("stderr", e.stderr.rstrip()) + "\n\n"
-
-        # Add guidance about automated fixes
-        if "files were modified by this hook" in e.stdout:
-            output += "Note: Some issues were automatically fixed by the pre-commit hooks. No manual fixes needed for those changes."
-        else:
-            output += "Note: The above issues require manual fixes as they were not automatically resolved."
-
-        return False, output.strip()
-    finally:
-        logger.info(
-            f"Pre-commit checks completed in {time.monotonic() - start_time:.2f}s"
-        )
-
-
-# Note: These functions provide core implementation that is called by the precommit and autocommit tools.
-# The tools provide the hook/command interface layer, while these functions provide the reusable core logic.
-def autocommit() -> Message:
-    """
-    Auto-commit changes made by gptme.
-
-    Returns a message asking the LLM to review changes and create a commit.
-    """
-    try:
-        # See if there are any changes to commit by checking for
-        # changes, excluding untracked files.
-        status_result_porcelain = subprocess.run(
-            ["git", "status", "--porcelain", "--untracked-files=no"],
-            capture_output=True,
-            text=True,
-            check=True,
-        )
-
-        if not status_result_porcelain.stdout.strip():
-            return Message("system", "No changes to commit.")
-
-        # Get current git status
-        status_result = subprocess.run(
-            ["git", "status"], capture_output=True, text=True, check=True
-        )
-
-        # Get git diff to show what changed
-        diff_result = subprocess.run(
-            ["git", "diff", "HEAD"], capture_output=True, text=True, check=True
-        )
-
-        # Create a message for the LLM to handle the commit
-        commit_prompt = f"""Pre-commit checks have passed and the following changes have been made:
-
-```git status
-{status_result.stdout}
-```
-
-```git diff HEAD
-{diff_result.stdout}
-```
-
-This is a good time to review these changes and consider creating an appropriate commit:
-
-1. Review the changes, decide which changes to include in the commit
-2. Stage only the relevant files using `git add` (never use `git add .` or `git add -A` to avoid adding unintended files)
-3. Create the commit using the HEREDOC format to avoid escaping issues. Both stage and commit in one go.
-
-```shell
-git add example.txt
-git commit -m "$(cat <<'EOF'
-Your commit message here
-EOF
-)"
-```
-"""
-
-        return Message("system", commit_prompt)
-
-    except subprocess.CalledProcessError as e:
-        return Message(
-            "system", f"Git operation failed: {e.stderr or e.stdout or str(e)}"
-        )
-    except Exception as e:
-        logger.error(f"Autocommit failed: {e}")
-        return Message("system", f"Autocommit failed: {e}")
 
 
 def enrich_messages_with_context(
